Change default join behavior with regard to null values
Previously, null values in the join key were considered a value like any other value. This meant that null values in the left frame would be joined with null values in the right frame. This is expensive and does not match default behavior in SQL.

Default behavior has now been changed to ignore null values in the join key. The previous behavior can be retained by setting join_nulls=True.